Move `df_to_arrow` to the io module
